webapp/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==2.5.2               # via kombu
ariadne==0.11.0           # via -r requirements.in
attrs==19.3.0             # via pytest
beautifulsoup4==4.7.1     # via -r requirements.in, django-htmlmin
billiard==3.6.3.0         # via celery
bleach==3.1.5             # via -r requirements.in
celery[redis]==4.4.2      # via -r requirements.in
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.9.2       # via social-auth-core
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-debug-toolbar==1.11  # via -r requirements.in
django-htmlmin==0.11.0    # via -r requirements.in
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.11.0       # via -r requirements.in
django-simple-sso==0.14.1  # via -r requirements.in
django==2.2.12            # via -r requirements.in, django-debug-toolbar, django-mptt, django-simple-sso, webservices
djangorestframework==3.9.4  # via -r requirements.in
faker==1.0.8              # via -r requirements.in
fastdiff==0.2.0           # via snapshottest
graphql-core==3.0.5       # via ariadne
html5lib==1.0.1           # via -r requirements.in, django-htmlmin
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via kombu, pluggy, pytest
itsdangerous==1.1.0       # via django-simple-sso, webservices
kombu==4.6.8              # via celery
markdown==2.6.11          # via -r requirements.in
more-itertools==8.3.0     # via pytest
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
packaging==20.4           # via bleach, pytest
pillow==6.2.2             # via -r requirements.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via -r requirements.in
py==1.8.1                 # via pytest
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via -r requirements.in, social-auth-core
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via faker
python3-openid==3.1.0     # via social-auth-core
pytz==2020.1              # via -r requirements.in, celery, django
redis==3.5.2              # via celery
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.23.0          # via -r requirements.in, requests-oauthlib, responses, social-auth-core, webservices
responses==0.10.14        # via -r requirements.in
six==1.15.0               # via bleach, cryptography, faker, html5lib, packaging, python-dateutil, responses, snapshottest, social-auth-app-django, social-auth-core
snapshottest==0.5.1       # via -r requirements.in
social-auth-app-django==3.1.0  # via -r requirements.in
social-auth-core==3.3.3   # via social-auth-app-django
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via django, django-debug-toolbar
starlette==0.13.4         # via ariadne
termcolor==1.1.0          # via snapshottest
text-unidecode==1.2       # via faker
typing-extensions==3.7.4.2  # via ariadne
unidecode==1.0.23         # via -r requirements.in
urllib3==1.25.9           # via requests
vine==1.3.0               # via amqp, celery
wasmer==0.4.1             # via fastdiff
wcwidth==0.1.9            # via pytest
webencodings==0.5.1       # via bleach, html5lib
webservices[django]==0.7  # via django-simple-sso
zipp==3.1.0               # via importlib-metadata

django-storages==1.6.6
zappa==0.51.0

git+https://${GIT_TOKEN}:@github.com/sleepio/core-framework-python.git@v1.8.10
git+https://${GIT_TOKEN}:@github.com/sleepio/bh-settings-python.git@v1.3.0
